ml/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
joblib
groq
python-dotenv